responsive tabpanel
