+++ package.json
@@ -14,6 +14,7 @@
     "@testing-library/react": "^11.2.3",
     "@testing-library/user-event": "^12.6.2",
     "babel-eslint": "^10.1.0",
+    "clsx": "^1.1.1",
     "dotenv": "^8.2.0",
     "emailjs-com": "^3.1.0",
     "eslint": "^7.27.0",
